Issue #94 Vollumfängliches Widget erzeugen, refactoring, fixing

+++ DroidsRocketColibri/src/ch/hsr/rocketcolibri/db/RocketColibriDB.java
@@ -56,6 +56,7 @@ public class RocketColibriDB {
 	}
 	
 	public void delete(Object obj){
+		if(obj==null)return;
 		tOdb.deleteCascade(obj);
 		tOdb.commit();
 	}
